requirements.txt
numpy
Pillow
future
pygame